refactor: add temp cache for config changes
We don't want to create new config changes with a ON CONFLICT DO NOTHING
clause because now we don't want the same config change to take up
multiple quotas from the rate limiter.

i.e. if an aws scraper is run @every 5m, we'll be trying to insert the
same config changes generated by the cloudtrail scraper again & again on
every run. The same change will take up one more quota from the rate
limiter on every run.

By knowing that the change already exist, we can avoid inserting that
change in the first place and the rate limiter will be happy about it.