fix: don't cache deleted configs
On the aws cluster, we had two HelmReleases (any many other duplicates)
under the same scraper with same KNN & external ids. One of them is
deleted & the other is active.

On cache init, we would fetch both the helm releases but the older
(deleted) one would replace the active one in the cache.